requirements/local.txt
# Local development requirements
-r base.txt

# Testing components
coverage==4.0

# debug toolbar
django-debug-toolbar==1.3.2